+++ app/src/main/java/com/example/a1l1/fragments/CitiesFragment.java
@@ -46,7 +46,6 @@ public class CitiesFragment extends Fragment {
             @Override
             public void onItemClick(String data) {
                 ((OnItemClick) requireActivity()).connect(data);
-                Snackbar.make(requireView(), "Вы выбрали город - " + data, Snackbar.LENGTH_LONG).show();
             }
         });
