+++ app/src/main/java/com/example/a1l1/fragments/WeatherFragment.java
@@ -11,9 +11,7 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import com.example.a1l1.R;
-import com.example.a1l1.models.WeatherRequest;
 
-import java.util.Locale;
 
 public class WeatherFragment extends Fragment {
     public final static String cityKey = "cityKey";
@@ -36,14 +34,18 @@ public class WeatherFragment extends Fragment {
 
         if (getArguments() != null) {
             String city = getArguments().getString(cityKey);
-            String degrees = getArguments().getString(degreesKey);
-            String pressure = getArguments().getString(pressureKey);
-            String humidity = getArguments().getString(humidityKey);
-            String windSpeed = getArguments().getString(windSpeedKey);
             cityNameView.setText(city);
-            degreesView.setText(degrees);
-            pressureView.setText(pressure);
-            humidityView.setText(humidity);
+
+            String temperatureValue = getArguments().getString(degreesKey);
+            degreesView.setText(temperatureValue);
+
+            String pressureText = getArguments().getString(pressureKey);
+            pressureView.setText(pressureText);
+
+            String humidityValue = getArguments().getString(humidityKey);
+            humidityView.setText(humidityValue);
+
+            String windSpeed = getArguments().getString(windSpeedKey);
             windSpeedView.setText(windSpeed);
         }
     }
@@ -54,21 +56,5 @@ public class WeatherFragment extends Fragment {
         pressureView = requireView().findViewById(R.id.PressureView);
         humidityView = requireView().findViewById(R.id.HumidityView);
         windSpeedView = requireView().findViewById(R.id.WindSpeedView);
-    }
-
-    public void updateCity(WeatherRequest weatherRequest) {
-        cityNameView.setText(weatherRequest.getName());
-
-        String temperatureValue = String.format(Locale.getDefault(), "%.2f", weatherRequest.getMain().getTemp());
-        degreesView.setText(temperatureValue);
-
-        String pressureText = String.format(Locale.getDefault(), "%d", weatherRequest.getMain().getPressure());
-        pressureView.setText(pressureText);
-
-        String waterValue = String.format(Locale.getDefault(), "%d", weatherRequest.getMain().getHumidity());
-        humidityView.setText(waterValue);
-
-        String windSpeed = String.format(Locale.getDefault(), " %.2f", weatherRequest.getWind().getSpeed());
-        windSpeedView.setText(windSpeed);
     }
 }
