+++ app/src/main/res/layout/activity_weather.xml
@@ -12,12 +12,13 @@
         android:layout_gravity="center_horizontal"
         android:layout_marginTop="15dp"
         android:layout_marginBottom="20dp"
+        android:textColor="@color/dark"
         android:textSize="40sp" />
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="50dp"
+        android:layout_marginTop="30dp"
         android:orientation="horizontal">
 
         <TextView
@@ -26,6 +27,7 @@
             android:layout_gravity="center_horizontal"
             android:layout_marginStart="105dp"
             android:text="@string/temp"
+            android:textColor="@color/dark"
             android:textSize="20sp" />
 
         <TextView
@@ -33,8 +35,10 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center_horizontal"
+            android:layout_marginStart="5dp"
             android:layout_marginTop="10dp"
             android:layout_marginBottom="10dp"
+            android:textColor="@color/dark"
             android:textSize="20sp" />
 
     </LinearLayout>
@@ -50,6 +54,7 @@
             android:layout_gravity="center_horizontal"
             android:layout_marginStart="105dp"
             android:text="@string/pressure_2"
+            android:textColor="@color/dark"
             android:textSize="20sp" />
 
         <TextView
@@ -57,8 +62,10 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center_horizontal"
+            android:layout_marginStart="5dp"
             android:layout_marginTop="10dp"
             android:layout_marginBottom="10dp"
+            android:textColor="@color/dark"
             android:textSize="20sp" />
 
     </LinearLayout>
@@ -74,6 +81,7 @@
             android:layout_gravity="center_horizontal"
             android:layout_marginStart="105dp"
             android:text="@string/humidity"
+            android:textColor="@color/dark"
             android:textSize="20sp" />
 
         <TextView
@@ -81,8 +89,10 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center_horizontal"
+            android:layout_marginStart="5dp"
             android:layout_marginTop="10dp"
             android:layout_marginBottom="10dp"
+            android:textColor="@color/dark"
             android:textSize="20sp" />
 
     </LinearLayout>
@@ -98,6 +108,7 @@
             android:layout_gravity="center_horizontal"
             android:layout_marginStart="105dp"
             android:text="@string/wind_speed2"
+            android:textColor="@color/dark"
             android:textSize="20sp" />
 
         <TextView
@@ -105,8 +116,8 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center_horizontal"
-            android:layout_marginTop="20dp"
-            android:layout_marginBottom="45dp"
+            android:layout_marginStart="5dp"
+            android:textColor="@color/dark"
             android:textSize="20sp" />
 
     </LinearLayout>
