+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">A1L1</string>
+    <string name="app_name">WeatherForecast</string>
     <string name="weather_forecast">Weather forecast</string>
     <string name="_2020">2020</string>
     <string name="enter_your_city">Enter your city</string>
